requirements.txt
Adafruit-BBIO==1.1.1
Adafruit-GPIO==1.0.3
Adafruit-Nokia-LCD==0.2.0
Adafruit-PureIO==0.2.3
automationhat==0.1.0
blinker==1.3
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
chardet==2.3.0
click==6.6
colorama==0.3.7
cryptography==1.7.1
drumhat==0.1.0
enum34==1.1.6
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==0.12.1
fourletterphat==0.1.0
future==0.17.1
get==2019.3.22
gpiozero==1.4.1
gyp==0.1
idna==2.2
ipaddress==1.0.17
itsdangerous==0.24
jabber-bot==0.1.1
Jinja2==2.8
keyring==10.1
keyrings.alt==1.3
MarkupSafe==0.23
microdotphat==0.2.1
mote==0.0.4
motephat==0.0.2
numpy==1.12.1
oauthlib==2.0.1
pantilthat==0.0.7
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
post==2019.3.22
psutil==5.6.1
public==2019.3.22
pyasn1==0.1.9
PyBluez==0.18
pycrypto==2.6.1
pygame==1.9.3
pygobject==3.22.0
pyinotify==0.9.6
PyJWT==1.4.2
pyOpenSSL==16.2.0
pyserial==3.2.1
python-twitter==3.5
pyudev==0.21.0
pyxdg==0.25
query-string==2019.3.22
rainbowhat==0.1.0
request==2019.3.22
requests==2.12.4
requests-oauthlib==0.7.0
RPi.GPIO==0.6.5
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-hat==2.2.0
simplejson==3.10.0
six==1.12.0
skywriter==0.0.7
sleekxmpp==1.3.3
sn3218==1.2.7
spidev==3.3
touchphat==0.0.1
twython==3.4.0
unicornhathd==0.0.4
urllib3==1.19.1
Werkzeug==0.11.15
wireless==0.3.2
xmpppy==0.5.0rc1
